Remove mojom dependency from all WebUIs
