+++ components/brave_rewards/resources/page/reducers/default_state.ts
@@ -6,6 +6,7 @@
 import {
   optional
 } from '../../../../brave_rewards/resources/shared/lib/optional'
+import * as Rewards from '../lib/types'
 
 export function defaultState (): Rewards.State {
   return {
