+++ components/brave_rewards/resources/page/reducers/index.ts
@@ -11,6 +11,7 @@ import rewardsReducer from './rewards_reducer'
 import walletReducer from './wallet_reducer'
 import promotionReducer from './promotion_reducer'
 import publishersReducer from './publishers_reducer'
+import * as Rewards from '../lib/types'
 
 function mergeReducers (state: Rewards.State | undefined, action: any) {
   if (!state) {
