+++ components/common/typescript.gni
@@ -23,7 +23,6 @@ brave_common_web_compile_inputs = [
   rebase_path("../definitions/newTab.d.ts"),
   rebase_path("../definitions/non-js-loaders.d.ts"),
   rebase_path("../definitions/prettier-bytes.d.ts"),
-  rebase_path("../definitions/rewards.d.ts"),
   rebase_path("../definitions/rewardsExtensions.d.ts"),
   rebase_path("../definitions/rewardsInternals.d.ts"),
   rebase_path("../definitions/throttleit.d.ts"),
